Added in working functions to fill a random integer array.
Instead of filling the integer arrays with random numbers, I am now filling them in order and then using a different function to then jumble up the array.
Added in a small function that will print the array to the console, this was done to save having to write in Writeline and Write commands when I want to see the results.