Added in loop to search for a magic square.
A while loop was added to continually search for a valid magic square.
Checks were also added when the user inputs a size to ensure the size entered is valid. If they enter a size that is greater than 3 they are reminded that it may take a while to find a solution and they are prompted to confirm whether they wish to continue.